Add graphics filenames to DB from .ods
Assumes that there are four different graphics, each with a description
in the .ods spreadsheet, that may appear in a different order at each
level. The descriptions at level 10 are used to automatically determine
the most common order of the graphics, which will have filenames marked
with the numbers 1, 2, 3, and 4. Then for the other levels, we check
whether the description for a graphic to show for a scene matches the
level 10 description, and use the level 10 number to determine the
filename (this is necessary if, e.g., at level 1, the level 10 scene 2
graphic should be shown, but as scene 1, since level 1 only has one
scene).