Change how graphics filenames are gotten from .ods
As it turns out, graphics may have different descriptions at different
levels -- there may not be only 4 graphics for a given story that are
shown in a different order. So we cannot use the descriptions to figure
out which graphic to load for which scene, since we don't know how they
will be tagged. Instead, there are now columns in each .ods story sheet
listing the tag of the graphic to load for each scene, for each level,
for each story. We use that tag plus the story name and level to
generate the graphic filename.

Also, update "stories_id" to "story_id" in all DB tables, since this id
will refer to a single story.