Receive level field with GameCommand.START message
When the game is sent a GameCommand.START message, a level to start
playing at can be provided. But since we track player progress
internally, we will ignore the provided level, except for printing out
an error if the level we pick based on our internal tracking is
different from the provided level.