Add <lookat_screen,b> before all robot responses
We want the robot to look at the screen as if considering a user
response before responding to it. So we add a "lookat_screen" action
command with the "blocking" flag to the start of each robot command that
will be sent after we receive a user response, so the robot will look at
the screen (and finish the lookat animation) before responding.