Fix pick next story script so one is always picked
In the scripts, the opal command LOAD_STORY was called before the STORY
command, which meant that the personalization manager had not
necessarily selected the next story when we wanted to load the story. We
have updated how the scripts are written to include a STORY SETUP
command that will just pick the next story, and added checks when
getting the story script or story details to ensure that a story has
been selected.