Move ros node initialization out of launch_game
For some reason, when the ros node was initialized within the
launch_game function, it never initialized. But when it was initialized
as a class variable, that was fine. So we've moved it back. The reason
to try putting it in the launch_game function was so you could try
running the game node and see the help text without having roscore
running, but looks like that won't work.