Use graphic names for order question answers
Previously, the answer options for order questions were "scene0",
"scene1", etc. But the names of the corresponding game objects loaded in
the SAR Opal game did not match this convention, so they were never
marked as correct or incorrect. The game objects loaded have names
matching the graphics file names. So we use these graphics names as the
names of the answer options for order questions.